FRY-36 Endpoint testing - Created API Endpoints + Restructured Folder Structure for URLs - Views missing - Added some commentary on available request methods for each endpoint
