--- Fridgify_Backend/fridge/fridge_urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-
-urlpatterns = [
-    path('', ),
-    path('management/', ),
-    path('item/', ),
-    path('content/', ),
-]
