+++ Fridgify_Backend/urls/authentication/auth_urls.py
@@ -1,8 +1,10 @@
-from django.urls import path
-
+from django.urls import path, include
 
 urlpatterns = [
+    # POST
     path('register/', ),
+    # POST
     path('login/', ),
+    # GET
     path('token/', ),
 ]
