+++ Fridgify_Backend/urls/fridge/content/content_urls.py
@@ -1,0 +1,8 @@
+from django.urls import path, include
+
+urlpatterns = [
+    # GET, POST
+    path('<int:fridge_id>/', ),
+    # GET, DELETE, POST
+    path('<int:fridge_id>/<int:item_id>', ),
+]
