+++ Fridgify_Backend/urls/fridge/item/item_urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+urlpatterns = [
+    # GET
+    path('', ),
+]
