+++ Fridgify_Backend/urls/fridge/management/management_urls.py
@@ -1,0 +1,12 @@
+from django.urls import path, include
+
+urlpatterns = [
+    # POST
+    path('', ),
+    # POST
+    path('join/', ),
+    # POST
+    path('create/', ),
+    # DELETE
+    path('<int:fridge_id>/', ),
+]
