+++ Fridgify_Backend/urls/stores/stores_urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+urlpatterns = [
+    # GET, POST
+    path('', ),
+]
